e2e: "vms manualsetup" builds auth'd server & client
These two Make targets stand up vagrant VMs representing a regular end-user
computer and a backup recipient (eg: raspberry pi). NOTE: the hope is they're
now setup in such a fashion as the rest of code committed for `make e2e`
represents the actual automation (much of the code committed here is trying to
mock setup that should probably be done manually in the real world).

Said targets are now a dependency of the e2e testing infrastructre i plan to use
for (see make target, "e2e"). No longer planning on unit tests, as I'm not sure
what those would be. Open to suggestions in the future.